requirements.txt
Pillow == 9.4.0
python-fsutil == 0.9.3
requests == 2.28.2
